Kleinigkeiten behoben
- Kleinere Fehler und Unschönheiten noch behoben und die App getestet mit diverse Kontakte erstellen, Geschenke erstellen, bearbeiten und archivieren. App hat ohne Fehler funktioniert. Kann nun als erste Version für geschlossenen Test im Google Play Store hochgeladen werden. Damit erste ausgewählte Benutzer die App weiter testen können und auf Fehler etc. prüfen können.

+++ lib/components/texts/centered_text.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class CenteredText extends StatelessWidget {
   final String text;
-  final int divider;
+  final double divider;
 
   const CenteredText({
     Key? key,
